lambdas/email/requirements.txt
boto3 
python-dotenv